+++ MortgageApi/Controllers/PodiumBaseController.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 using PodiumInterview.MortgageApi.Models;
-using Swashbuckle.Swagger.Annotations;
+using Swashbuckle.AspNetCore.Annotations;
 
 namespace PodiumInterview.MortgageApi.Controllers
 {
@@ -18,7 +18,7 @@ namespace PodiumInterview.MortgageApi.Controllers
         /// including any <see cref="ModelStateDictionary">ModelState</see> errors.
         /// </summary>
         [ApiExplorerSettings(IgnoreApi = true)]
-        [SwaggerResponse(HttpStatusCode.BadRequest, Type = typeof(ApiErrorResult))]
+        [SwaggerResponse((int)HttpStatusCode.BadRequest, Type = typeof(ApiErrorResult))]
         public IActionResult ApiValidationError(IEnumerable<string> extraErrors = null)
         {
             //Construct errors from ModelState
@@ -37,7 +37,7 @@ namespace PodiumInterview.MortgageApi.Controllers
         /// Returns a "500 Internal Server Error" API error.
         /// </summary>
         [ApiExplorerSettings(IgnoreApi = true)]
-        [SwaggerResponse(HttpStatusCode.InternalServerError, Type = typeof(ApiErrorResult))]
+        [SwaggerResponse((int)HttpStatusCode.InternalServerError, Type = typeof(ApiErrorResult))]
         public IActionResult ApiServerError(object additionalInfo = null)
         {
             return StatusCode(StatusCodes.Status500InternalServerError, additionalInfo);
